Reorganize wifi module
