+++ esp-wifi/src/wifi/event.rs
@@ -1,8 +1,115 @@
 use alloc::boxed::Box;
 
+use enumset::EnumSetType;
 use esp_hal::sync::Locked;
+use num_derive::FromPrimitive;
 
-use super::WifiEvent;
+/// Events generated by the WiFi driver.
+#[repr(i32)]
+#[derive(Debug, FromPrimitive, EnumSetType)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum WifiEvent {
+    /// Wi-Fi is ready for operation.
+    WifiReady = 0,
+    /// Scan operation has completed.
+    ScanDone,
+    /// Station mode started.
+    StaStart,
+    /// Station mode stopped.
+    StaStop,
+    /// Station connected to a network.
+    StaConnected,
+    /// Station disconnected from a network.
+    StaDisconnected,
+    /// Station authentication mode changed.
+    StaAuthmodeChange,
+
+    /// Station WPS succeeds in enrollee mode.
+    StaWpsErSuccess,
+    /// Station WPS fails in enrollee mode.
+    StaWpsErFailed,
+    /// Station WPS timeout in enrollee mode.
+    StaWpsErTimeout,
+    /// Station WPS pin code in enrollee mode.
+    StaWpsErPin,
+    /// Station WPS overlap in enrollee mode.
+    StaWpsErPbcOverlap,
+
+    /// Soft-AP start.
+    ApStart,
+    /// Soft-AP stop.
+    ApStop,
+    /// A station connected to Soft-AP.
+    ApStaconnected,
+    /// A station disconnected from Soft-AP.
+    ApStadisconnected,
+    /// Received probe request packet in Soft-AP interface.
+    ApProbereqrecved,
+
+    /// Received report of FTM procedure.
+    FtmReport,
+
+    /// AP's RSSI crossed configured threshold.
+    StaBssRssiLow,
+    /// Status indication of Action Tx operation.
+    ActionTxStatus,
+    /// Remain-on-Channel operation complete.
+    RocDone,
+
+    /// Station beacon timeout.
+    StaBeaconTimeout,
+
+    /// Connectionless module wake interval has started.
+    ConnectionlessModuleWakeIntervalStart,
+
+    /// Soft-AP WPS succeeded in registrar mode.
+    ApWpsRgSuccess,
+    /// Soft-AP WPS failed in registrar mode.
+    ApWpsRgFailed,
+    /// Soft-AP WPS timed out in registrar mode.
+    ApWpsRgTimeout,
+    /// Soft-AP WPS pin code in registrar mode.
+    ApWpsRgPin,
+    /// Soft-AP WPS overlap in registrar mode.
+    ApWpsRgPbcOverlap,
+
+    /// iTWT setup.
+    ItwtSetup,
+    /// iTWT teardown.
+    ItwtTeardown,
+    /// iTWT probe.
+    ItwtProbe,
+    /// iTWT suspended.
+    ItwtSuspend,
+    /// TWT wakeup event.
+    TwtWakeup,
+    /// bTWT setup.
+    BtwtSetup,
+    /// bTWT teardown.
+    BtwtTeardown,
+
+    /// NAN (Neighbor Awareness Networking) discovery has started.
+    NanStarted,
+    /// NAN discovery has stopped.
+    NanStopped,
+    /// NAN service discovery match found.
+    NanSvcMatch,
+    /// Replied to a NAN peer with service discovery match.
+    NanReplied,
+    /// Received a follow-up message in NAN.
+    NanReceive,
+    /// Received NDP (Neighbor Discovery Protocol) request from a NAN peer.
+    NdpIndication,
+    /// NDP confirm indication.
+    NdpConfirm,
+    /// NAN datapath terminated indication.
+    NdpTerminated,
+    /// Wi-Fi home channel change, doesn't occur when scanning.
+    HomeChannelChange,
+
+    /// Received Neighbor Report response.
+    StaNeighborRep,
+}
 
 pub(crate) mod sealed {
     use super::*;
@@ -15,6 +122,7 @@ pub(crate) mod sealed {
         unsafe fn from_raw_event_data(ptr: *mut crate::binary::c_types::c_void) -> Self;
     }
 }
+
 /// The type of handlers of events.
 pub type Handler<T> = dyn FnMut(&T) + Sync + Send;
