+++ examples/src/bin/wifi_access_point.rs
@@ -29,9 +29,8 @@ use esp_println::{print, println};
 use esp_wifi::{
     init,
     wifi::{
+        config::{AccessPointConfiguration, Configuration},
         event::{self, EventExt},
-        AccessPointConfiguration,
-        Configuration,
     },
 };
 use smoltcp::iface::{SocketSet, SocketStorage};
