+++ examples/src/bin/wifi_embassy_access_point.rs
@@ -32,11 +32,10 @@ use esp_println::{print, println};
 use esp_wifi::{
     init,
     wifi::{
-        AccessPointConfiguration,
-        Configuration,
+        config::{AccessPointConfiguration, Configuration},
+        event::WifiEvent,
         WifiController,
         WifiDevice,
-        WifiEvent,
         WifiState,
     },
     EspWifiController,
