+++ examples/src/bin/wifi_embassy_access_point_with_sta.rs
@@ -40,12 +40,10 @@ use esp_println::{print, println};
 use esp_wifi::{
     init,
     wifi::{
-        AccessPointConfiguration,
-        ClientConfiguration,
-        Configuration,
+        config::{AccessPointConfiguration, ClientConfiguration, Configuration},
+        event::WifiEvent,
         WifiController,
         WifiDevice,
-        WifiEvent,
         WifiState,
     },
     EspWifiController,
