+++ examples/src/bin/wifi_sniffer.rs
@@ -46,7 +46,7 @@ fn main() -> ! {
     let (mut controller, interfaces) =
         esp_wifi::wifi::new(&esp_wifi_ctrl, peripherals.WIFI).unwrap();
 
-    controller.set_mode(wifi::WifiMode::Sta).unwrap();
+    controller.set_mode(wifi::config::WifiMode::Sta).unwrap();
     controller.start().unwrap();
 
     let mut sniffer = interfaces.sniffer;
