+++ examples/src/bin/wifi_static_ip.rs
@@ -27,7 +27,10 @@ use esp_hal::{
 use esp_println::{print, println};
 use esp_wifi::{
     init,
-    wifi::{AccessPointInfo, ClientConfiguration, Configuration, WifiError},
+    wifi::{
+        config::{AccessPointInfo, ClientConfiguration, Configuration},
+        WifiError,
+    },
 };
 use smoltcp::iface::{SocketSet, SocketStorage};
